requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.1.3
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.3.0
tqdm==4.67.1
urllib3==2.2.3
wheel==0.43.0